// src/modules/webinars/enum/webinars.enum.ts
export enum WebinarCategory {
  LEGACY = 'LEGACY',
  THINK_TANK_BOARD = 'THINK_TANK_BOARD',
  BRANDS_RECTORAL = 'BRANDS_RECTORAL',
  ROAD_MAP_RECTORAL = 'ROAD_MAP_RECTORAL',
}

export enum MonthsOfYear {
  JANUARY = 'JANUARY',
  FEBRUARY = 'FEBRUARY',
  MARCH = 'MARCH',
  APRIL = 'APRIL',
  MAY = 'MAY',
  JUNE = 'JUNE',
  JULY = 'JULY',
  AUGUST = 'AUGUST',
  SEPTEMBER = 'SEPTEMBER',
  OCTOBER = 'OCTOBER',
  NOVEMBER = 'NOVEMBER',
  DECEMBER = 'DECEMBER',
}